+++ src/Oro/Bundle/PimDataGridBundle/Repository/DatagridViewRepository.php
@@ -3,8 +3,9 @@
 namespace Oro\Bundle\PimDataGridBundle\Repository;
 
 use Akeneo\UserManagement\Component\Model\UserInterface;
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
 use Doctrine\DBAL\Connection;
-use Doctrine\DBAL\FetchMode;
 use Doctrine\ORM\EntityRepository;
 use Oro\Bundle\PimDataGridBundle\Entity\DatagridView;
 
@@ -34,7 +35,7 @@ SQL;
             'owner_id' => $user->getId(),
         ]);
 
-        return $statement->fetchAll(FetchMode::COLUMN);
+        return $statement->fetchAll(\PDO::FETCH_COLUMN);
     }
 
     /**
@@ -45,7 +46,7 @@ SQL;
         string $alias,
         string $term = '',
         array $options = []
-    ): array {
+    ): Collection {
         $options += ['limit' => 20, 'page' => 1];
         $offset = (int) $options['limit'] * ((int) $options['page'] - 1);
 
@@ -71,7 +72,7 @@ SQL;
             $qb->setParameter('ids', $identifiers);
         }
 
-        return $qb->getQuery()->execute();
+        return new ArrayCollection($qb->getQuery()->execute());
     }
 
     private function getConnection(): Connection